Assignment_2/pyproject.toml
[project]
name = "class-assignment-1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chainlit>=2.6.0",
    "openai-agents>=0.1.0",
    "python-decouple>=3.8",
]
